+++ CHANGELOG.rst
@@ -4,7 +4,9 @@ History
 
 master (XXXX-XX-XX)
 --------------------
+
 * Feature: Add support for the optional 'networks' parameter on server creation.
+* Feature: Add python 3.9 support
 
 1.6.3 (2020-01-09)
 --------------------